requirements.txt
newspaper3k
streamlit
google-generativeai
python-dotenv
fpdf
gunicorn